+++ types/interfaces/definitions.ts
@@ -61,6 +61,7 @@ export { default as parachains } from './parachains/definitions.ts';
 export { default as poll } from './poll/definitions.ts';
 export { default as purchase } from './purchase/definitions.ts';
 
+export { default as dryRunApi } from './dryRunApi/definitions.ts';
 export { default as xcm } from './xcm/definitions.ts';
 export { default as xcmPaymentApi } from './xcmPaymentApi/definitions.ts';
 export { default as xcmRuntimeApi } from './xcmRuntimeApi/definitions.ts';